+++ src/pages/FlowRun/FlowRun.vue
@@ -4,7 +4,7 @@ import { mapGetters } from 'vuex'
 import Actions from '@/pages/FlowRun/Actions'
 import BreadCrumbs from '@/components/BreadCrumbs'
 import DetailsTile from '@/pages/FlowRun/Details-Tile'
-import FlowRunPageGanttChart from '@/pages/FlowRunPageGanttChart'
+import TimelineTile from '@/pages/FlowRun/Timeline-Tile'
 import LogsCard from '@/components/LogsCard/LogsCard'
 import SchematicTile from '@/pages/FlowRun/Schematic-Tile'
 import SubPageNav from '@/layouts/SubPageNav'
@@ -12,6 +12,7 @@ import TaskRunHeartbeatTile from '@/pages/FlowRun/TaskRunHeartbeat-Tile'
 import TaskRunTableTile from '@/pages/FlowRun/TaskRunTable-Tile'
 import TileLayout from '@/layouts/TileLayout'
 import TileLayoutFull from '@/layouts/TileLayout-Full'
+import { parser } from '@/utils/markdownParser'
 
 export default {
   metaInfo() {
@@ -35,7 +36,7 @@ export default {
     Actions,
     BreadCrumbs,
     DetailsTile,
-    FlowRunPageGanttChart,
+    TimelineTile,
     LogsCard,
     SchematicTile,
     SubPageNav,
@@ -52,6 +53,9 @@ export default {
   computed: {
     ...mapGetters('tenant', ['tenant']),
     ...mapGetters('api', ['isCloud']),
+    flowId() {
+      return this.flowRun?.flow_id
+    },
     flowRunId() {
       return this.$route.params.id
     },
@@ -75,6 +79,9 @@ export default {
         case 'chart':
           query = { chart: '' }
           break
+        case 'artifacts':
+          query = { artifacts: '' }
+          break
         default:
           break
       }
@@ -90,7 +97,11 @@ export default {
       if ('schematic' in this.$route.query) return 'schematic'
       if ('logId' in this.$route.query) return 'logs'
       if ('chart' in this.$route.query) return 'chart'
+      if ('artifacts' in this.$route.query) return 'artifacts'
       return 'overview'
+    },
+    parseMarkdown(md) {
+      return parser(md)
     }
   },
   apollo: {
@@ -161,18 +172,41 @@ export default {
         Schematic
       </v-tab>
 
-      <v-tab href="#chart" :style="hideOnMobile">
-        <v-icon left>pi-gantt</v-icon>
-        Gantt Chart
-      </v-tab>
-
       <v-tab href="#logs" :style="hideOnMobile" data-cy="flow-run-logs-tab">
         <v-icon left>format_align_left</v-icon>
         Logs
       </v-tab>
 
-      <v-tab-item class="tab-full-height pa-0" value="overview">
+      <v-tab href="#artifacts" :style="hideOnMobile" disabled>
+        <v-badge color="codePink" content="Coming Soon!" bottom bordered inline>
+          <v-icon left>fas fa-fingerprint</v-icon>
+          Artifacts
+        </v-badge>
+      </v-tab>
+    </v-tabs>
+
+    <v-tabs-items
+      v-model="tab"
+      class="px-6 mx-auto tabs-border-bottom"
+      :style="{ 'max-width': tab == 'chart' ? 'auto' : '1440px' }"
+      mandatory
+    >
+      <v-tab-item
+        class="tab-full-height pa-0"
+        value="overview"
+        transition="quick-fade"
+        reverse-transition="quick-fade"
+      >
         <TileLayout>
+          <TimelineTile
+            v-if="flowId"
+            slot="row-0"
+            condensed
+            :flow-id="flowId"
+            :flow-run-id="flowRunId"
+            :flow-run="flowRun"
+          />
+
           <DetailsTile slot="row-2-col-1-row-1-tile-1" :flow-run="flowRun" />
 
           <TaskRunHeartbeatTile
@@ -187,19 +221,23 @@ export default {
         </TileLayout>
       </v-tab-item>
 
-      <v-tab-item class="tab-full-height" value="schematic">
+      <v-tab-item
+        class="tab-full-height"
+        value="schematic"
+        transition="quick-fade"
+        reverse-transition="quick-fade"
+      >
         <TileLayoutFull>
           <SchematicTile slot="row-2-tile" />
         </TileLayoutFull>
       </v-tab-item>
 
-      <v-tab-item class="tab-full-height" value="chart">
-        <v-card class="pa-2 mt-2" tile>
-          <FlowRunPageGanttChart :flow-run-id="$route.params.id" />
-        </v-card>
-      </v-tab-item>
-
-      <v-tab-item class="tab-full-height" value="logs">
+      <v-tab-item
+        class="tab-full-height"
+        value="logs"
+        transition="quick-fade"
+        reverse-transition="quick-fade"
+      >
         <TileLayoutFull>
           <LogsCard
             slot="row-2-tile"
@@ -214,7 +252,11 @@ export default {
           />
         </TileLayoutFull>
       </v-tab-item>
-    </v-tabs>
+
+      <v-tab-item class="tab-full-height" value="artifacts">
+        <!-- <div v-html="parseMarkdown('# Hello!')"></div> -->
+      </v-tab-item>
+    </v-tabs-items>
 
     <v-bottom-navigation v-if="$vuetify.breakpoint.smAndDown" fixed>
       <v-btn @click="tab = 'overview'">
@@ -227,14 +269,14 @@ export default {
         <v-icon>pi-schematic</v-icon>
       </v-btn>
 
-      <v-btn @click="tab = 'chart'">
-        Gantt Chart
-        <v-icon>pi-gantt</v-icon>
-      </v-btn>
-
       <v-btn @click="tab = 'logs'">
         Logs
         <v-icon>format_align_left</v-icon>
+      </v-btn>
+
+      <v-btn disabled @click="tab = 'artifacts'">
+        Artifacts
+        <v-icon>fas fa-fingerprint</v-icon>
       </v-btn>
     </v-bottom-navigation>
   </v-sheet>
@@ -247,5 +289,16 @@ export default {
 
 .tab-full-height {
   min-height: 80vh;
+}
+
+/* stylelint-disable */
+
+.v-tab--disabled .v-badge__badge {
+  pointer-events: none;
+}
+
+.v-badge--inline .v-badge__badge,
+.v-badge--inline .v-badge__wrapper {
+  margin: 5px;
 }
 </style>